+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 # Change Log
 All notable changes to this project will be documented in this file. This change log follows the conventions of [keepachangelog.com](http://keepachangelog.com/).
 
+## Unreleased
+
+### Changed
+- `defsingleton` now takes an (optional) options map with an optional `:reload?` flag.
+
 ## [0.1.1]
 
 ### Fixed